separate implementation of album chat from preview so that the preview can be used by itself in the album page
