+++ src/components/album/AlbumPreview.js
@@ -1,24 +1,24 @@
 import React, { useEffect, useState } from 'react';
 import { CardContent } from '@mui/material';
-import AlbumChat from '../album/AlbumChat';
 
 const AlbumPreview = ({
-  apiUrl,
+  spotifyUrl,
   style = {},
   wide = false,
   width = wide ? '100%' : 300,
   height = wide ? 80 : 380,
   frameBorder = 0,
   allow = 'encrypted-media',
+  children,
   ...props
 }) => {
-  const [spotifyUrl, setSpotifyUrl] = useState('');
+  const [spotifyEmbedUrl, setSpotifyEmbedUrl] = useState('');
 
   useEffect(() => {
-    // regular expression pattern to extract the album ID from the API URL
+    // regular expression pattern to extract the album ID from the Spotify **API** Url that's saved on the album object
     const regexPattern = /\/albums\/([a-zA-Z0-9]+)/;
     //  extract the album ID
-    const match = apiUrl.match(regexPattern);
+    const match = spotifyUrl.match(regexPattern);
     if (!match) {
       // if URL doesn't match the expected format
       console.error('Invalid API URL');
@@ -26,16 +26,16 @@ const AlbumPreview = ({
     }
     //extract the album id
     const albumId = match[1];
-    //  Spotify album URL
+    //  Spotify album URL for embedding
     const spotifyAlbumUrl = `https://open.spotify.com/embed/album/${albumId}`;
-    setSpotifyUrl(spotifyAlbumUrl);
-  }, [apiUrl]);
+    setSpotifyEmbedUrl(spotifyAlbumUrl);
+  }, [spotifyUrl]);
 
   return (
     <CardContent style={{ marginTop: '1rem', width: '100%', display: 'flex' }}>
       <iframe
         title="Spotify Web Player"
-        src={spotifyUrl}
+        src={spotifyEmbedUrl}
         width={width}
         height={height}
         frameBorder={frameBorder}
@@ -43,9 +43,9 @@ const AlbumPreview = ({
         allow={allow}
         {...props}
       />
-      <div style={{ width: '100%', display: 'flex' }}>
-        <AlbumChat apiUrl={apiUrl} />
-      </div>
+      {children && (
+        <div style={{ width: '100%', display: 'flex' }}>{children}</div>
+      )}
     </CardContent>
   );
 };
